# packages/contracts/openapi/ragdoll.openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentDetail": {
        "properties": {
          "chunk_count": {
            "minimum": 0.0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_size": {
            "minimum": 0.0,
            "title": "File Size",
            "type": "integer"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "indexed_chunk_count": {
            "minimum": 0.0,
            "title": "Indexed Chunk Count",
            "type": "integer"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "original_text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text Content"
          },
          "preview_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Text"
          },
          "processing_status": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "source_kind": {
            "enum": [
              "manual_upload",
              "external_sync"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploaded_by": {
            "format": "uuid",
            "title": "Uploaded By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "uploaded_by",
          "title",
          "original_filename",
          "mime_type",
          "file_type",
          "file_size",
          "source_kind",
          "processing_status",
          "chunk_count",
          "indexed_chunk_count",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
...
      "DocumentUpdateRequest": {
        "properties": {
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "space_id"
        ],
        "title": "DocumentUpdateRequest",
        "type": "object"
      },
      "EntityDetailResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "document_count": {
            "minimum": 0.0,
            "title": "Document Count",
            "type": "integer"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "graph_node_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Node Id"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/EntityHistoryEntry"
            },
            "title": "History",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_mentioned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Mentioned At"
          },
          "mention_count": {
            "minimum": 0.0,
            "title": "Mention Count",
            "type": "integer"
          },
          "normalized_name": {
            "title": "Normalized Name",
            "type": "string"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/EntityMentionRecord"
            },
            "title": "Provenance",
            "type": "array"
          },
          "related_documents": {
            "items": {
              "$ref": "#/components/schemas/EntityRelatedDocument"
            },
            "title": "Related Documents",
            "type": "array"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "entity_type",
          "display_name",
          "normalized_name",
          "mention_count",
          "document_count",
          "created_at",
          "updated_at"
        ],
        "title": "EntityDetailResponse",
        "type": "object"
      },
      "EntityHistoryEntry": {
        "properties": {
          "citation": {
            "$ref": "#/components/schemas/Citation"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "mention_id": {
            "format": "uuid",
            "title": "Mention Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "surface_text": {
            "title": "Surface Text",
            "type": "string"
          }
        },
        "required": [
          "mention_id",
          "document_id",
          "surface_text",
          "observed_at",
          "citation"
        ],
        "title": "EntityHistoryEntry",
        "type": "object"
      },
...
      "EntityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "EntityListResponse",
        "type": "object"
      },
...
      "GraphLinkResponse": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "relation_type": {
            "title": "Relation Type",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "source_id",
          "target_id",
          "relation_type",
          "weight"
        ],
        "title": "GraphLinkResponse",
        "type": "object"
      },
...
      "GraphResponse": {
        "properties": {
          "depth": {
            "minimum": 1.0,
            "title": "Depth",
            "type": "integer"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/GraphLinkResponse"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeResponse"
            },
            "title": "Nodes",
            "type": "array"
          },
          "seed_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Entity Id"
          }
        },
        "required": [
          "depth"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
...
      "MutationResult": {
        "description": "Generic mutation response envelope.",
        "example": {
          "message": "Updated successfully.",
          "success": true
        },
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "MutationResult",
        "type": "object"
      },
...
      "ProblemResponse": {
        "description": "Stable problem+json response body.",
        "example": {
          "code": "authentication_required",
          "detail": "Authentication credentials are required for this resource.",
          "instance": "/api/v1/spaces",
          "status": 401,
          "title": "Authentication required",
          "type": "https://ragdoll.dev/problems/authentication-required"
        },
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "default": "about:blank",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "default": "HTTP request failed",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "https://ragdoll.dev/problems/http-error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "status",
          "detail"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
      "ProcessingStatus": {
        "description": "Stage-aware processing status contract.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "extraction": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          },
          "graph": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          },
          "overall": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          },
          "parsing": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          },
          "upload": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          },
          "vector": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProcessingStageStatus"
              }
            ],
            "default": "pending"
          }
        },
        "title": "ProcessingStatus",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Ragdoll API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_route_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested document was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Document metadata exists but backing storage is unavailable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document Route",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "read_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested document was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Document metadata exists but backing storage is unavailable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Document",
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "patch_document_api_v1_documents__document_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested document was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Document metadata exists but backing storage is unavailable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/entities": {
      "get": {
        "operationId": "read_entities_api_v1_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          },
          {
            "in": "query",
            "name": "all_spaces",
            "required": false,
            "schema": {
              "default": false,
              "title": "All Spaces",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested entity was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Entities",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/v1/entities/{entity_id}": {
      "get": {
        "operationId": "read_entity_detail_api_v1_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          },
          {
            "in": "query",
            "name": "all_spaces",
            "required": false,
            "schema": {
              "default": false,
              "title": "All Spaces",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested entity was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Entity Detail",
        "tags": [
          "entities"
        ]
      }
    },
...
    "/api/v1/knowledge-graph/documents/{document_id}": {
      "get": {
        "operationId": "read_document_graph_api_v1_knowledge_graph_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          },
          {
            "in": "query",
            "name": "all_spaces",
            "required": false,
            "schema": {
              "default": false,
              "title": "All Spaces",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested graph seed was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Document Graph",
        "tags": [
          "knowledge_graph"
        ]
      }
    },
    "/api/v1/knowledge-graph/entities/{entity_id}/subgraph": {
      "get": {
        "operationId": "read_entity_subgraph_api_v1_knowledge_graph_entities__entity_id__subgraph_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          },
          {
            "in": "query",
            "name": "all_spaces",
            "required": false,
            "schema": {
              "default": false,
              "title": "All Spaces",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Requested graph seed was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Entity Subgraph",
        "tags": [
          "knowledge_graph"
        ]
      }
    },
...
  }
}